bugfix: fix auth delimiters

+++ public/index.php
@@ -18,22 +18,11 @@ $db = new Phalcon\Db\Adapter\Pdo\Sqlite($config);
 $request = new Phalcon\Http\Request();
 $response = new Phalcon\Http\Response();
 $shell = new shell(false);
+$res = false;
 
-$app->before(function() use ($app, $db, $request, $response)
+$app->before(function() use ($db, $request, $response)
 {
-    if($request->getServer('PHP_AUTH_USER'))
-    {
-        $apiKey = $request->getServer('PHP_AUTH_USER');
-        $res = $db->fetchOne('SELECT * FROM `api` WHERE `key` = :key LIMIT 1', Phalcon\Db::FETCH_ASSOC, array('key' => $apiKey));
-        if(!$res)
-        {
-            utils::send401($response);
-        }
-    }
-    else
-    {
-        utils::send401($response);
-    }
+    utils::authCheck($db, $request, $response);
 });
 
 $app->get('/api/v1/{task}/{target}[/]?{mask}', function($task, $target, $mask = null) use ($app, $db, $response, $request, $shell)
@@ -140,28 +129,31 @@ $app->get('/api/v1/{task}/{target}[/]?{mask}', function($task, $target, $mask = 
     return $response;
 });
 
-$app->get('/api/v1/stream/{uuid}', function($uuid) use ($app, $db, $response)
+$app->get('/api/v1/stream/{uuid}', function($uuid) use ($app, $db, $response, $res)
 {
-    $data = $db->fetchOne("SELECT * FROM `streams` WHERE `uuid` = :uuid LIMIT 1", Phalcon\Db::FETCH_ASSOC, array('uuid' => $uuid));
-    if(!$data)
-        utils::send403($response);
-    else
+    //needed because we're not using Phalcon\HTTP\Request() for this route
+    if (utils::$auth)
     {
-        $ip = long2ip($data['target']);
-        $type = $data['type'];
-        $init = new shell(true);
-        switch($type)
-        {
-            case 'ipv4':
-                $init->execute('traceroute -w 1', '-A', $ip);
-            break;
-            case 'ipv6':
-                $init->execute('traceroute6 -w 1', '-A', $ip);
-            break;
+        $data = $db->fetchOne("SELECT * FROM `streams` WHERE `uuid` = :uuid LIMIT 1", Phalcon\Db::FETCH_ASSOC, array('uuid' => $uuid));
+        if (!$data)
+            utils::send403($response);
+        else {
+            $ip = long2ip($data['target']);
+            $type = $data['type'];
+            $init = new shell(true);
+            switch ($type) {
+                case 'ipv4':
+                    $init->execute('traceroute -w 1', '-A', $ip);
+                    break;
+                case 'ipv6':
+                    $init->execute('traceroute6 -w 1', '-A', $ip);
+                    break;
+            }
+            $db->execute("DELETE FROM `streams` WHERE `uuid` = ?", array($uuid));
         }
-        $db->execute("DELETE FROM `streams` WHERE `uuid` = ?", array($uuid));
     }
 });
+
 
 $app->put('/api/v1/update-key/{key}', function($key) use ($app, $response, $db)
 {
